requirements.txt
marshmallow>=3
flask==2.2.2
flask-restful
flask-migrate
flask-sqlalchemy
flask-marshmallow
flask-jwt-extended
marshmallow-sqlalchemy
python-dotenv
passlib
apispec[yaml]
apispec-webframeworks
tox
gunicorn
psycopg2-binary